Add new `format=ids` result format for auto-suggest ID insertion
Returns a list of JSON maps with `label` and `value` fields. The
labels are the possible completions of the term, the values are
the corresponding IDs.

When used with the jQuery-UI auto-suggest widget, it will suggest
the labels and insert the values when a label is selected. For UI
elements with other requirements the result will have to be
processed according to their needs.

The auto-suggest usage sample on the main page now uses this result
format. See also the `format` column in the API endpoints table for
documentation (now contains links for supported result formats).