Change internal data format to `expanded` JSON-LD
Triggered by the update to Elasticsearch 0.90.5 (see #110), which
complains about an inconsistency in our previous data format:

With `comapct` JSON-LD, plain literals are serialized as strings,
while e.g. literals with a language are serialized as an array of
objects with `@value` and `@lang` keys. This results in different
types for the same field (here: string and array).

Since Elasticsearch requires that a specific field has a consistent
type, this commit changes the JSON-LD serialization to use the
`expanded` format, which always encodes literals as object arrays.

At the same time, the API still serves the `compact` format, so
clients should not have to change their result processing logic.

As the format of the data created by the conversion from N-Triples
to JSON-LD has changed, this commit increments the version number
of lodmill-ld to 0.2.0-SNAPSHOT.

It also deletes the Eclipse project config files in lodmill-ui,
which have to be generated locally with `play eclipse`.