Remove RDF type from properties to use for detail resolution (#317)
The type of the main entity is always included, since that triple
has the main entity as subject. The resolved entities are all
included for their labels. Their types cause issues when
restricting queries by type. Since a solution that retains these
types would be quite involved, and we have no requirement for them
to be included, we simply remove them from the resolution config.