Support restricting queries by type (see #317)
Uses `t` as the query parameter. The obvious choice of `type` is
currently not possible due to a bug in the Play framework [1], so
we will set up a rule in our Apache proxy to rewrite `type` to `t`
until the issue is resolved. We can then switch from `t` to `type`
internally and remove the rewrite, without affecting API clients.

[1] https://github.com/playframework/playframework/issues/2259