Add parameter to get unmodified elasticsearch _source data
To get the productive api data a lookup must be made using the http-api,
not the elasticsearch directly. The latter doesn't work because
the elasticsearch node of the api is not exposed to the world
and thus travis cannot get the data.
Since the data which is to be indexed to make tests must *exectly*
look like the data which was imported into the elasticsearch backend
the new format parameter "internal" is introduced to serve the unmodified
_source field residing in elasticsearch documents.