Add collection "NwBib" to test data
* add workaround for getting the data of resource/NwBib

As there is no extra endpoint "collection" this data (e.g. "NwBib")
is subsumed under the "resource" endpoint. But as the collection
data is NOT stored in the "resource" index, API query parameters
will not work. Nothing will work except the path style of the API.
So this is a workaround for getting the collection data. Not having
the untouched "_source" data from elasticsearch is ok because the
only functionality tested is using the path style (and no elaborated
query (like the id-query) which would demand the elaborated structure
residing under "_source").