Raise parameter "size" from 100 to 1000
Without the cache it's safe to get more data because this data
isn't stored anymore. The pro of getting more data is a cosiderable
less HTTP overhead when serving the data, e.g. when using the
SampleUsage class.

The limit in SampleUsage is set to 300. This seems to be the fastest
value, increasing getting the data up to 30% (this' only roughly).
While not threaded, higher values results in a longer time
where either elasticsearch has nothing to do (because the API is
crunching the result) or it's the API waiting to get the results
from Elasticsearch.
Takes around 100 minutes to get the 360k nwbib documents.