Revert to use the productive API
With the resolving of the memory leak (see #107) and is save
to revert f28703ba536ffe2b1c64801067fe95ed2e83347c to use the
production API.
While the test API may be restarted multiple times a day this would
break the usage of this program if it's running at the same time.
So the usage of the production API will be much better.