Update email address, remove lobid-only comm channels
As the page is about the team, we should not provide
links to GitHub and mailing list that only apply to
lobid as one of many services the group provides.

This is basically another step of moving forward
from being "lobid team" to being "Offene Infrastruktur"
or as of now "Metadateninfrastruktur".