Use "jsonId" instead of "id" (#501)
The "id" is used internally by gatsby and is not the value of the json field
"id" but rather an internally generated id. So the value of your json's "id"
is stored in "jsonId". Use this field to access the json's "id".

See https://www.gatsbyjs.com/plugins/gatsby-transformer-json/#id-and-jsonid-key.