Fix ids in project, publication and product queries
With #503 Gatsby did not find any more the correct id for the publications and so on.
Because the gatsby-transformer-plugin changes `id` to jsonId`, but only on the top level of nodes things are a bit messed up.
Unfortunatley it does this automatically, and it is not configurable (https://www.gatsbyjs.com/plugins/gatsby-transformer-json/#id-and-jsonid-key, otherwise we could for example always use jsonId).